From Tom Norris on 2022-09-18:
Attached is a file containing items numbered 200-299 along with their description, if there is one. I’ve added placeholder numbers for those item nbrs that were not in the original spreadsheet.

The description needs to be cleaned up as we discussed by comparing it to the PDF file. Also, the descriptions have a “* “ in front which needs to be deleted.

From the spreadsheet you gave me, there does not appear to be any titles or descriptions for any items numbered 300+. I assume this is because the student interns were unable to finish their work before CRMII shutdown.